edited Settings.tsx
